@@ -5,57 +5,76 @@ publish_to: "none"
 version: 1.0.1+2
 
 environment:
-  sdk: ">=2.18.5 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   animated_splash_screen: ^1.3.0
-  camera: ^0.10.0+4
-  cupertino_icons: ^1.0.5
+  camera: ^0.10.4
   double_back_to_close_app: ^2.1.0
-  easy_folder_picker: ^1.3.0
-  flutter_launcher_icons: ^0.10.0
+  easy_folder_picker: ^1.3.3
+  flutter_launcher_icons: ^0.13.1
   flutter_spinkit: ^5.1.0
-  flutter_svg: ^1.1.6
-  fluttertoast: ^8.1.1
+  flutter_svg: ^2.0.5
+  fluttertoast: ^8.2.1
   gallery_saver: ^2.3.2
   hive: ^2.2.3
   hive_flutter: ^1.1.0
-  image_cropper: ^3.0.1
-  image_picker: ^0.8.6
-  intl: ^0.17.0
-  introduction_screen: ^3.1.1
-  open_file: ^3.2.1
-  package_info_plus: ^1.4.2
+  http: ^0.13.0
+  image: ^4.0.17
+  image_cropper: ^3.0.3
+  image_picker: ^0.8.7+4
+  intl: ^0.19.0
+  introduction_screen: ^3.1.8
+  open_file: ^3.3.1
+  package_info_plus: ^3.1.2
   path: ^1.8.2
-  path_provider: ^2.0.11
+  path_provider: ^2.0.14
   pdf: ^3.8.4
   pdf_viewer_plugin: ^2.0.1
-  permission_handler: ^10.2.0
-  photofilters: ^3.0.1
-  printing: ^5.9.3
-  provider: ^6.0.4
-  quick_actions: ^1.0.1
+  permission_handler: ^11.3.1
+  photofilters:
+    git: https://github.com/jinosh05/photofilters
+  printing: ^5.13.1
+  provider: ^6.0.5
+  quick_actions: ^1.0.2
   rating_dialog: ^2.0.4
   # share: ^2.0.4
-  share_plus: ^4.5.3
-  shared_preferences: ^2.0.15
-  source_span: ^1.9.0
-  url_launcher: ^6.1.7
-  wiredash: ^1.5.0
+  share_plus: ^6.3.4
+  shared_preferences: ^2.1.0
+  source_span: ^1.10.0
+  url_launcher: ^6.1.10
+  wiredash: ^2.2.0
 
 dev_dependencies:
   build_runner: ^2.3.2
-  flutter_test:
-    sdk: flutter
+  # flutter_test:
+  #   sdk: flutter
   hive_generator: ^2.0.0
-lint: ^1.10.0
+  flutter_native_splash: ^2.2.16
+lint: ^2.1.2
 
 flutter_icons:
   android: true
   ios: true
   image_path: "assets/app_icon.png"
+
+flutter_native_splash:
+  color: "#ffffff"
+  image: assets/app_icon.png
+  branding: assets/doclense.png
+  color_dark: "000000"
+  image_dark: assets/app_icon.png
+  branding_dark: assets/doclense.png
+
+  android_12:
+    image: assets/app_icon.png
+    icon_background_color: "#ffffff"
+    image_dark: assets/app_icon.png
+    icon_background_color_dark: "#000000"
+
+  web: false
 
 flutter:
   uses-material-design: true
