@@ -1,33 +1,31 @@
-import 'dart:io';
+import "dart:io";
 
-import 'package:doclense/configs/app_dimensions.dart';
-import 'package:doclense/configs/app_typography.dart';
-import 'package:doclense/configs/space.dart';
-import 'package:doclense/configs/ui.dart';
-import 'package:doclense/constants/appstrings.dart';
-import 'package:doclense/constants/route_constants.dart';
-import 'package:doclense/ui_components/grid_item.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_spinkit/flutter_spinkit.dart';
-import 'package:gallery_saver/gallery_saver.dart';
-import 'package:image_picker/image_picker.dart';
-
-import '../providers/image_list.dart';
+import "package:doclense/configs/app_dimensions.dart";
+import "package:doclense/configs/app_typography.dart";
+import "package:doclense/configs/space.dart";
+import "package:doclense/configs/ui.dart";
+import "package:doclense/constants/appstrings.dart";
+import "package:doclense/constants/route_constants.dart";
+import "package:doclense/providers/image_list.dart";
+import "package:doclense/ui_components/grid_item.dart";
+import "package:flutter/material.dart";
+import "package:flutter_spinkit/flutter_spinkit.dart";
+import "package:gallery_saver/gallery_saver.dart";
+import "package:image_picker/image_picker.dart";
 
 class MultiDelete extends StatefulWidget {
+  const MultiDelete(this.imageList, {super.key});
   final ImageList imageList;
 
-  const MultiDelete(this.imageList);
-
   @override
-  _MultiDeleteState createState() => _MultiDeleteState();
+  MultiDeleteState createState() => MultiDeleteState();
 }
 
-class _MultiDeleteState extends State<MultiDelete> {
+class MultiDeleteState extends State<MultiDelete> {
   List<Item>? itemList;
   List<Item>? selectedList;
   File? imageFile;
-  final picker = ImagePicker();
+  final ImagePicker picker = ImagePicker();
 
   bool _isLoading = true;
 
@@ -35,219 +33,216 @@ class _MultiDeleteState extends State<MultiDelete> {
   void initState() {
     loadList();
     super.initState();
-    Future.delayed(
-        const Duration(seconds: 1),
-        () => setState(() {
-              _isLoading = false;
-            }));
   }
 
   void loadList() {
-    itemList = [];
-    selectedList = [];
+    itemList = <Item>[];
+    selectedList = <Item>[];
     for (int i = 0; i < (widget.imageList.length()); i++) {
       itemList?.add(Item(widget.imageList.imagelist.elementAt(i), i));
     }
+    setState(() {
+      _isLoading = false;
+    });
   }
 
-  Future<void> _showChoiceDialogDel(BuildContext context) {
-    return showDialog(
+  Future<void> _showChoiceDialogDel(BuildContext context) => showDialog(
         context: context,
-        builder: (BuildContext context) {
-          return AlertDialog(
-            backgroundColor: Colors.blueGrey[800],
-            title: Text(
-              S.deleteSelected,
-              textAlign: TextAlign.center,
-              style: AppText.b1!.cl(Colors.white),
-            ),
-            content: SingleChildScrollView(
-              child: ListBody(
-                children: <Widget>[
-                  GestureDetector(
-                    onTap: () {
-                      setState(() {
-                        for (int i = 0; i < selectedList!.length; i++) {
-                          final int idx = widget.imageList.imagelist.indexOf(
-                              itemList![itemList!.indexOf(selectedList![i])]
-                                  .imageUrl);
-                          widget.imageList.imagelist.removeAt(idx);
-                          widget.imageList.imagepath.removeAt(idx);
-                          itemList!.remove(selectedList![i]);
-                        }
-                        selectedList = [];
-                      });
-                      Navigator.pop(context);
-                    },
-                    child: Text(
-                      S.yes,
-                      textAlign: TextAlign.center,
-                      style: AppText.b1!.cl(Colors.white),
-                    ),
-                  ),
-                  Space.y!,
-                  GestureDetector(
-                    onTap: () {
-                      Navigator.of(context).pop();
-                    },
-                    child: Text(
-                      S.no,
-                      textAlign: TextAlign.center,
-                      style: AppText.b1!.cl(Colors.white),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          );
-        });
-  }
-
-  Future<void> _showChoiceDialogHome(BuildContext context) {
-    return showDialog(
-        context: context,
-        builder: (BuildContext ctx) {
-          return AlertDialog(
-            backgroundColor: Colors.blueGrey[800],
-            title: Text(
-              S.deleteWarning,
-              textAlign: TextAlign.center,
-              style: AppText.b1!.cl(Colors.white),
-            ),
-            content: SingleChildScrollView(
-              child: ListBody(
-                children: <Widget>[
-                  GestureDetector(
-                    onTap: () {
-                      for (int i = 0; i < itemList!.length; i++) {
-                        print('i = $i');
-                        print(widget.imageList.imagelist.length);
+        builder: (BuildContext context) => AlertDialog(
+          backgroundColor: Colors.blueGrey[800],
+          title: Text(
+            S.deleteSelected,
+            textAlign: TextAlign.center,
+            style: AppText.b1!.cl(Colors.white),
+          ),
+          content: SingleChildScrollView(
+            child: ListBody(
+              children: <Widget>[
+                GestureDetector(
+                  onTap: () {
+                    setState(() {
+                      for (int i = 0; i < selectedList!.length; i++) {
                         final int idx = widget.imageList.imagelist.indexOf(
-                            itemList![itemList!.indexOf(itemList![i])]
-                                .imageUrl);
+                          itemList![itemList!.indexOf(selectedList![i])]
+                              .imageUrl,
+                        );
                         widget.imageList.imagelist.removeAt(idx);
                         widget.imageList.imagepath.removeAt(idx);
-                        // itemList.remove(idx);
-                        itemList!.removeAt(idx);
+                        itemList!.remove(selectedList![i]);
                       }
-
-                      Navigator.of(context).popUntil((route) => route.isFirst);
-                    },
-                    child: Text(
-                      S.yes,
-                      textAlign: TextAlign.center,
-                      style: AppText.b1!.cl(Colors.white),
-                    ),
+                      selectedList = <Item>[];
+                    });
+                    Navigator.pop(context);
+                  },
+                  child: Text(
+                    S.yes,
+                    textAlign: TextAlign.center,
+                    style: AppText.b1!.cl(Colors.white),
                   ),
-                  Space.y!,
-                  GestureDetector(
-                    onTap: () {
-                      Navigator.of(ctx).pop();
-                    },
-                    child: Text(
-                      S.no,
-                      textAlign: TextAlign.center,
-                      style: AppText.b1!.cl(Colors.white),
-                    ),
+                ),
+                Space.y!,
+                GestureDetector(
+                  onTap: () {
+                    Navigator.of(context).pop();
+                  },
+                  child: Text(
+                    S.no,
+                    textAlign: TextAlign.center,
+                    style: AppText.b1!.cl(Colors.white),
                   ),
-                ],
-              ),
+                ),
+              ],
             ),
-          );
-        });
-  }
+          ),
+        ),
+      );
+
+  Future<void> _showChoiceDialogHomeScreen(BuildContext context) => showDialog(
+        context: context,
+        builder: (BuildContext ctx) => AlertDialog(
+          backgroundColor: Colors.blueGrey[800],
+          title: Text(
+            S.deleteWarning,
+            textAlign: TextAlign.center,
+            style: AppText.b1!.cl(Colors.white),
+          ),
+          content: SingleChildScrollView(
+            child: ListBody(
+              children: <Widget>[
+                GestureDetector(
+                  onTap: () {
+                    for (int i = 0; i < itemList!.length; i++) {
+                      debugPrint("i = $i");
+                      debugPrint(widget.imageList.imagelist.length.toString());
+                      final int idx = widget.imageList.imagelist.indexOf(
+                        itemList![itemList!.indexOf(itemList![i])].imageUrl,
+                      );
+                      widget.imageList.imagelist.removeAt(idx);
+                      widget.imageList.imagepath.removeAt(idx);
+                      // itemList.remove(idx);
+                      itemList!.removeAt(idx);
+                    }
+
+                    Navigator.of(context)
+                        .popUntil((Route<dynamic> route) => route.isFirst);
+                  },
+                  child: Text(
+                    S.yes,
+                    textAlign: TextAlign.center,
+                    style: AppText.b1!.cl(Colors.white),
+                  ),
+                ),
+                Space.y!,
+                GestureDetector(
+                  onTap: () {
+                    Navigator.of(ctx).pop();
+                  },
+                  child: Text(
+                    S.no,
+                    textAlign: TextAlign.center,
+                    style: AppText.b1!.cl(Colors.white),
+                  ),
+                ),
+              ],
+            ),
+          ),
+        ),
+      );
 
   Future<void> _openGallery() async {
-    XFile? picture = await picker.pickImage(source: ImageSource.gallery);
+    final XFile? picture = await picker.pickImage(source: ImageSource.gallery);
     setState(() {
-      if (picture != null) imageFile = File(picture.path);
+      if (picture != null) {
+        imageFile = File(picture.path);
+      }
     });
     if (imageFile != null) {
-      Navigator.of(context).pushNamed(
+      // ignore: use_build_context_synchronously
+      await Navigator.of(context).pushNamed(
         RouteConstants.imageView,
-        arguments: {
-          'imageFile': imageFile,
-          'imageList': widget.imageList,
+        arguments: <String, Object?>{
+          "imageFile": imageFile,
+          "imageList": widget.imageList,
         },
       );
     }
   }
 
   Future<void> _openCamera() async {
-    XFile? picture = await picker.pickImage(source: ImageSource.camera);
+    final XFile? picture = await picker.pickImage(source: ImageSource.camera);
     setState(() {
-      if (picture != null) imageFile = File(picture.path);
+      if (picture != null) {
+        imageFile = File(picture.path);
+      }
     });
 
     if (imageFile != null) {
-      GallerySaver.saveImage(imageFile!.path)
-          .then((value) => print("Image Saved"));
-      Navigator.of(context).pushNamed(
-        RouteConstants.imageView,
-        arguments: {
-          'imageFile': imageFile,
-          'imageList': widget.imageList,
-        },
-      );
+      await GallerySaver.saveImage(imageFile!.path)
+          .then((bool? value) => debugPrint("Image Saved"))
+          .whenComplete(() async {
+        await Navigator.of(context).pushNamed(
+          RouteConstants.imageView,
+          arguments: <String, Object?>{
+            "imageFile": imageFile,
+            "imageList": widget.imageList,
+          },
+        );
+      });
     }
   }
 
-  Future<void> _showChoiceDialogAdd(BuildContext context) {
-    return showDialog(
+  Future<void> _showChoiceDialogAdd(BuildContext context) => showDialog(
         context: context,
-        builder: (BuildContext context) {
-          return AlertDialog(
-            backgroundColor: Colors.blueGrey[800],
-            title: Text(
-              S.addMorePages,
-              textAlign: TextAlign.center,
-              style: AppText.b1!.cl(Colors.white),
-            ),
-            content: SingleChildScrollView(
-              child: ListBody(
-                children: <Widget>[
-                  GestureDetector(
-                    onTap: () {
-                      Navigator.of(context).pop();
-                      _openGallery();
-                    },
-                    child: Text(
-                      S.gallery,
-                      style: AppText.b1!.cl(Colors.white),
-                    ),
+        builder: (BuildContext context) => AlertDialog(
+          backgroundColor: Colors.blueGrey[800],
+          title: Text(
+            S.addMorePages,
+            textAlign: TextAlign.center,
+            style: AppText.b1!.cl(Colors.white),
+          ),
+          content: SingleChildScrollView(
+            child: ListBody(
+              children: <Widget>[
+                GestureDetector(
+                  onTap: () {
+                    Navigator.of(context).pop();
+                    _openGallery();
+                  },
+                  child: Text(
+                    S.gallery,
+                    style: AppText.b1!.cl(Colors.white),
                   ),
-                  Space.y!,
-                  GestureDetector(
-                    onTap: () {
-                      Navigator.of(context).pop();
-                      _openCamera();
-                    },
-                    child: Text(
-                      S.camera,
-                      style: AppText.b1!.cl(Colors.white),
-                    ),
+                ),
+                Space.y!,
+                GestureDetector(
+                  onTap: () {
+                    Navigator.of(context).pop();
+                    _openCamera();
+                  },
+                  child: Text(
+                    S.camera,
+                    style: AppText.b1!.cl(Colors.white),
                   ),
-                ],
-              ),
+                ),
+              ],
             ),
-          );
-        });
-  }
+          ),
+        ),
+      );
 
   @override
-  Widget build(BuildContext context) {
-    return WillPopScope(
-      onWillPop: () async {
-        if (itemList!.isNotEmpty) {
-          setState(() {
-            widget.imageList.imagelist.removeAt(itemList!.length - 1);
-            widget.imageList.imagepath.removeAt(itemList!.length - 1);
-            itemList!.removeAt(itemList!.length - 1);
-          });
-        }
-        return true;
-      },
-      child: Scaffold(
+  Widget build(BuildContext context) => WillPopScope(
+        onWillPop: () async {
+          if (itemList!.isNotEmpty) {
+            setState(() {
+              widget.imageList.imagelist.removeAt(itemList!.length - 1);
+              widget.imageList.imagepath.removeAt(itemList!.length - 1);
+              itemList!.removeAt(itemList!.length - 1);
+            });
+          }
+          return true;
+        },
+        child: Scaffold(
           // backgroundColor: Colors.blueGrey[100],
           appBar: getAppBar(),
           body: _isLoading
@@ -257,92 +252,90 @@ class _MultiDeleteState extends State<MultiDelete> {
               : GridView.builder(
                   itemCount: itemList!.length,
                   gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
-                      crossAxisCount: UI.width! >= 640 ? 3 : 2,
-                      crossAxisSpacing: 4,
-                      mainAxisSpacing: 4),
-                  itemBuilder: (context, index) {
-                    return Card(
-                      elevation: 10,
-                      child: GridItem(
-                          item: itemList![index],
-                          isSelected: (bool value) {
-                            setState(() {
-                              if (value) {
-                                selectedList!.add(itemList![index]);
-                              } else {
-                                selectedList!.remove(itemList![index]);
-                              }
-                            });
-                            print("$index : $value");
-                          },
-                          key: Key(itemList![index].rank.toString())),
-                    );
-                  }),
+                    crossAxisCount: UI.width! >= 640 ? 3 : 2,
+                    crossAxisSpacing: 4,
+                    mainAxisSpacing: 4,
+                  ),
+                  itemBuilder: (BuildContext context, int index) => Card(
+                    elevation: 10,
+                    child: GridItem(
+                      item: itemList![index],
+                      isSelected: (bool value) {
+                        setState(() {
+                          if (value) {
+                            selectedList!.add(itemList![index]);
+                          } else {
+                            selectedList!.remove(itemList![index]);
+                          }
+                        });
+                        debugPrint("$index : $value");
+                      },
+                      key: Key(itemList![index].rank.toString()),
+                    ),
+                  ),
+                ),
           floatingActionButton: FloatingActionButton(
-            onPressed: () {},
-            child: IconButton(
-              iconSize: AppDimensions.font(18),
-              onPressed: () {
-                _showChoiceDialogAdd(context);
-              },
-              icon: const Icon(
-                Icons.add,
-              ),
+            onPressed: () async {
+              await _showChoiceDialogAdd(context);
+            },
+            child: Icon(
+              Icons.add,
+              size: AppDimensions.font(18),
             ),
-          )),
-    );
-  }
+          ),
+        ),
+      );
 
-  AppBar getAppBar() {
-    return AppBar(
-      leading: IconButton(
-        onPressed: () {
-          if (itemList!.isNotEmpty) {
-            setState(() {
-              widget.imageList.imagelist.removeAt(itemList!.length - 1);
-              widget.imageList.imagepath.removeAt(itemList!.length - 1);
-              itemList!.removeAt(itemList!.length - 1);
-            });
-          }
-          Navigator.of(context).pop();
-        },
-        icon: const Icon(Icons.arrow_back),
-      ),
-      title: Text(
-        selectedList!.isEmpty
-            ? S.documents
-            : "${selectedList!.length} item selected",
-      ),
-      actions: <Widget>[
-        if (selectedList!.isEmpty)
-          Container()
-        else
-          IconButton(
+  AppBar getAppBar() => AppBar(
+        leading: IconButton(
+          onPressed: () {
+            if (itemList!.isNotEmpty) {
+              setState(() {
+                widget.imageList.imagelist.removeAt(itemList!.length - 1);
+                widget.imageList.imagepath.removeAt(itemList!.length - 1);
+                itemList!.removeAt(itemList!.length - 1);
+              });
+            }
+            Navigator.of(context).pop();
+          },
+          icon: const Icon(Icons.arrow_back),
+        ),
+        title: Text(
+          selectedList!.isEmpty
+              ? S.documents
+              : "${selectedList!.length} item selected",
+        ),
+        actions: <Widget>[
+          if (selectedList!.isEmpty)
+            Container()
+          else
+            IconButton(
               icon: const Icon(Icons.delete),
-              onPressed: () {
-                _showChoiceDialogDel(context);
-              }),
-        IconButton(
+              onPressed: () async {
+                await _showChoiceDialogDel(context);
+              },
+            ),
+          IconButton(
             icon: const Icon(Icons.picture_as_pdf),
-            onPressed: () {
-              Navigator.of(context).pushNamed(
+            onPressed: () async {
+              await Navigator.of(context).pushNamed(
                 RouteConstants.pdfConversionScreen,
                 arguments: widget.imageList,
               );
-            }),
-        IconButton(
+            },
+          ),
+          IconButton(
             icon: const Icon(Icons.home),
-            onPressed: () {
-              _showChoiceDialogHome(context);
-            })
-      ],
-    );
-  }
+            onPressed: () async {
+              await _showChoiceDialogHomeScreen(context);
+            },
+          ),
+        ],
+      );
 }
 
 class Item {
+  Item(this.imageUrl, this.rank);
   File imageUrl;
   int rank;
-
-  Item(this.imageUrl, this.rank);
 }
