@@ -1,6 +1,5 @@
-import 'package:flutter/material.dart';
-import 'dart:ui' as ui;
-import 'ui.dart';
+import "package:doclense/configs/ui.dart";
+import "package:flutter/material.dart";
 
 class AppDimensions {
   static double? maxContainerWidth;
@@ -12,10 +11,10 @@ class AppDimensions {
 
   static Size? size;
 
-  static init() {
+  static void init() {
     ratio = UI.width! / UI.height!;
-    double pixelDensity = UI.mediaQuery().devicePixelRatio;
-    ratio = (ratio) + ((pixelDensity + ratio) / 2);
+    final double pixelDensity = UI.mediaQuery().devicePixelRatio;
+    ratio = ratio + ((pixelDensity + ratio) / 2);
 
     if (UI.width! <= 380 && pixelDensity >= 3) {
       ratio *= 0.85;
@@ -27,8 +26,8 @@ class AppDimensions {
     padding = ratio * 3;
   }
 
-  static _initLargeScreens() {
-    const safe = 2.4;
+  static void _initLargeScreens() {
+    const double safe = 2.4;
 
     ratio *= 1.5;
 
@@ -37,7 +36,7 @@ class AppDimensions {
     }
   }
 
-  static _initSmallScreensHighDensity() {
+  static void _initSmallScreensHighDensity() {
     if (!UI.sm! && ratio > 2.0) {
       ratio = 2.0;
     }
@@ -50,8 +49,8 @@ class AppDimensions {
   }
 
   static String inString() {
-    final x = UI.width! / UI.height!;
-    final ps = ui.window.physicalSize;
+    final double x = UI.width! / UI.height!;
+    final Size ps = UI.physicalSize!;
     return """
       Width: ${UI.width} | ${ps.width}
       Height: ${UI.height} | ${ps.height}
@@ -62,30 +61,27 @@ class AppDimensions {
   }
 
   /// Use this method for creating responsive Space between widgets
-  static double space([double multiplier = 1.0]) {
-    return AppDimensions.padding! * 3 * multiplier;
-  }
+  static double space([double multiplier = 1.0]) =>
+      AppDimensions.padding! * 3 * multiplier;
 
   /// Use this method for creating responsive Widget Sizes
   /// Note: For Exact width and height never use it
-  static double normalize(double unit) {
-    return (AppDimensions.ratio * unit * 0.77) + unit;
-  }
+  static double normalize(double unit) =>
+      (AppDimensions.ratio * unit * 0.77) + unit;
 
   /// Use this method for creating responsive Fonts
-  static double font(double unit) {
-    return (AppDimensions.ratio * unit * 0.125) + unit * 1.90;
-  }
+  static double font(double unit) =>
+      (AppDimensions.ratio * unit * 0.125) + unit * 1.90;
 
   /// Use this method for getting responsive width
   /// Note: 4.8 is just for handling failure case
-  static double width(double percentage) {
-    return (UI.horizontal ?? 4.8) * percentage;
-  }
+  static double width(double percentage) => (UI.horizontal ?? 4.8) * percentage;
 
   /// Use this method for getting responsive height
   /// Note: 7.2 is just for handling failure case
-  static double height(double percentage) {
-    return (UI.vertical ?? 7.2) * percentage;
-  }
+  static double height(double percentage) => (UI.vertical ?? 7.2) * percentage;
+
+  /// For Responsive Circular Border Radius
+  static BorderRadius borRadius(double radius) =>
+      BorderRadius.circular(normalize(radius));
 }
