@@ -1,6 +1,5 @@
-import 'package:flutter/material.dart';
-
-import 'app_dimensions.dart';
+import "package:doclense/configs/app_dimensions.dart";
+import "package:flutter/material.dart";
 
 class AppText {
   static TextStyle? btn;
@@ -27,10 +26,10 @@ class AppText {
   static TextStyle? l2;
   static TextStyle? l2b;
 
-  static init() {
-    const b = FontWeight.bold;
+  static void init() {
+    const FontWeight b = FontWeight.bold;
 
-    const baseStyle = TextStyle(//fontFamily:
+    const TextStyle baseStyle = TextStyle(//fontFamily:
         );
 
     h1 = baseStyle.copyWith(fontSize: AppDimensions.font(22));
@@ -60,23 +59,17 @@ class AppText {
 }
 
 extension SuperTextStyles on TextStyle {
-  TextStyle cl(Color color) {
-    return copyWith(
-      color: color,
-    );
-  }
+  TextStyle cl(Color color) => copyWith(
+        color: color,
+      );
 
-  TextStyle s([double? size]) {
-    return copyWith(
-      fontSize: size ?? fontSize,
-    );
-  }
+  TextStyle s([double? size]) => copyWith(
+        fontSize: size ?? fontSize,
+      );
 
-  TextStyle tsc([double multiplier = 1.0]) {
-    return copyWith(
-      fontSize: fontSize! * multiplier,
-    );
-  }
+  TextStyle tsc([double multiplier = 1.0]) => copyWith(
+        fontSize: fontSize! * multiplier,
+      );
 
   TextStyle w(int i) {
     FontWeight w = FontWeight.w400;
