@@ -1,17 +1,21 @@
-import 'dart:math';
-import 'package:flutter/widgets.dart';
+import "dart:math";
+
+import "package:flutter/widgets.dart";
 
 class UI {
-  static MediaQueryData? _mediaQueryData;
+  static late MediaQueryData? _mediaQueryData;
   static double? width;
   static double? height;
   static double? horizontal;
   static double? vertical;
-  static EdgeInsets? padding;
-  static EdgeInsets? vi;
 
-  static double? _safeAreaHorizontal;
-  static double? _safeAreaVertical;
+  /// Padding Before Keyboard raise
+  static EdgeInsets? padding;
+
+  /// Padding After Keyboard raise
+  static EdgeInsets? vi;
+  static late double? _safeAreaHorizontal;
+  static late double? _safeAreaVertical;
   static double? safeWidth;
   static double? safeHeight;
 
@@ -26,6 +30,9 @@ class UI {
   static bool? xl;
   static bool? xlg;
   static bool? xxlg;
+
+  static bool isPortrait = false;
+  static Size? physicalSize;
 
   static void init(BuildContext context) {
     _mediaQueryData = MediaQuery.of(context);
@@ -42,12 +49,16 @@ class UI {
         _mediaQueryData!.padding.left + _mediaQueryData!.padding.right;
     _safeAreaVertical =
         _mediaQueryData!.padding.top + _mediaQueryData!.padding.bottom;
-    safeWidth = (width! - _safeAreaHorizontal!);
-    safeHeight = (height! - _safeAreaVertical!);
+    safeWidth = width! - _safeAreaHorizontal!;
+    safeHeight = height! - _safeAreaVertical!;
+
+    ///
+    isPortrait = (width ?? 0) < 600;
+    physicalSize = View.of(context).physicalSize;
   }
 
-  static initChecks(MediaQueryData query) {
-    var size = query.size;
+  static void initChecks(MediaQueryData query) {
+    final Size size = query.size;
     diagonal = sqrt((size.width * size.width) + (size.height * size.height));
     xxs = size.width > 300;
     xs = size.width > 360;
