@@ -1,119 +1,115 @@
 /// S is for AppStrings
 class S {
-  static final String next = 'Next';
+  static const String next = "Next";
 
-  static final String done = 'Done';
+  static const String done = "Done";
 
-  static final String skip = 'Skip';
+  static const String skip = "Skip";
 
-  static final String scanDocuments =
-      'Scan Your Favourite Documents or Assignments on the go!!';
+  static const String scanDocuments =
+      "Scan Your Favourite Documents or Assignments on the go!!";
 
-  static final String appMadeInIndia = 'An App made in 🇮🇳 with ❤️';
+  static const String appMadeInIndia = "An App made in 🇮🇳 with ❤️";
 
-  static final String team = 'Team';
+  static const String team = "Team";
 
-  static final String presentingTeam = "Presenting The Team";
+  static const String presentingTeam = "Presenting The Team";
 
-  static final String developers = 'Developers';
+  static const String developers = "Developers";
 
-  static final String contributors = 'Contributors';
+  static const String contributors = "Contributors";
 
-  static final String preferences = 'Preferences';
+  static const String preferences = "Preferences";
 
-  static final String appThemeOptions = 'App theme options';
+  static const String appThemeOptions = "App theme options";
 
-  static final String darkMode = 'Dark Mode';
+  static const String darkMode = "Dark Mode";
 
-  static final String reportABug = 'Report a bug';
+  static const String reportABug = "Report a bug";
 
-  static final String requestAFeature = 'Request A Feature';
+  static const String requestAFeature = "Request A Feature";
 
-  static final String contactDevelopers = 'Contact Developers';
+  static const String contactDevelopers = "Contact Developers";
 
-  static final String aboutApp = "About App";
+  static const String aboutApp = "About App";
 
-  static final String appAbout =
-      "DocLense is the one place for all your documents!\n\nYou can now click, upload, crop, rotate and do so\nmuch more!\n\nSo whether it is your college assignment or the\noffice document you want to digitalize, stop\nworrying and just use";
+  static const String appAbout = """
+DocLense is the one place for all your documents!\n\nYou can now click, upload, crop, rotate and do so\nmuch more!\n\nSo whether it is your college assignment or the\noffice document you want to digitalize, stop\nworrying and just use""";
 
-  static final String doclense = 'Doclense';
+  static const String doclense = "Doclense";
 
-  static final String madeByOpenSource = "Made with ❤ by Open Source";
+  static const String madeByOpenSource = "Made with ❤ by Open Source";
 
-  static final String noPDFyet = "No PDFs Scanned Yet !! ";
+  static const String noPDFyet = "No PDFs Scanned Yet !! ";
 
-  static final String yourPDF = 'Your PDF!';
+  static const String yourPDF = "Your PDF!";
 
-  static final String deleteConfirmText =
+  static const String deleteConfirmText =
       "The PDF will be permanently deleted.\nDo you want to proceed?";
 
-  static final String yes = 'Yes';
+  static const String yes = "Yes";
 
-  static final String no = 'No';
+  static const String no = "No";
 
-  static final String rename = 'Rename';
+  static const String rename = "Rename";
 
-  static final String save = 'Save';
+  static const String save = "Save";
 
-  static final String removeStar = 'Removed from starred documents';
+  static const String removeStar = "Removed from starred documents";
 
-  static final String addStar = 'Added to starred documents!';
+  static const String addStar = "Added to starred documents!";
 
-  static final String pdfPreview = "Pdf preview";
+  static const String pdfPreview = "Pdf preview";
 
-  static final String starredDocs = 'Starred Docs';
+  static const String starredDocs = "Starred Docs";
 
-  static final String back = 'Back';
+  static const String back = "Back";
 
-  static final String undo = 'Undo';
+  static const String undo = "Undo";
 
-  static final String crop = 'Crop';
+  static const String crop = "Crop";
 
-  static final String filter = 'Filter';
+  static const String filter = "Filter";
 
-  static final String deleteWarning =
+  static const String deleteWarning =
       "All your progress will be lost.\nDo you want to go back to home?";
 
-  static final String onePlace = "One Place For All \n Your Documents!";
+  static const String onePlace = "One Place For All \n Your Documents!";
 
-  static final String home = 'Home';
+  static const String home = "Home";
 
-  static final String starredDocuments = 'Starred Documents';
+  static const String starredDocuments = "Starred Documents";
 
-  static final String settings = 'Settings';
+  static const String settings = "Settings";
 
-  static final String rateus = 'Rate us';
+  static const String rateus = "Rate us";
 
-  static final String shareTheApp = 'Share the App';
+  static const String shareTheApp = "Share the App";
 
-  static final String shareMessage =
-      'Hey !! , I am using this wonderful app : DocLense , check it out here https://github.com/smaranjitghose/DocLense';
+  static const String shareMessage =
+      "Hey !! , I am using this wonderful app : DocLense , check it out here https://github.com/smaranjitghose/DocLense";
 
-  static final String howWasExperience = "How was your experience?";
+  static const String howWasExperience = "How was your experience?";
 
-  static final String submit = 'Submit';
+  static const String submit = "Submit";
 
-  static final String letUsKnow = "Let us know what you think";
+  static const String letUsKnow = "Let us know what you think";
 
-  static final String deleteSelected = 'Delete selected item(s)?';
+  static const String deleteSelected = "Delete selected item(s)?";
 
-  static final String addMorePages = 'Add more pages with:';
+  static const String addMorePages = "Add more pages with:";
 
-  static final String gallery = 'Gallery';
+  static const String gallery = "Gallery";
 
-  static final String camera = 'Camera';
+  static const String camera = "Camera";
 
-  static final String documents = "Documents";
+  static const String documents = "Documents";
 
-  static final String nameYourPdf = 'Name Your PDF';
+  static const String nameYourPdf = "Name Your PDF";
 
-  static final String pdfName = 'Pdf Name';
-
-  // static final String a = '';
+  static const String pdfName = "Pdf Name";
 
   // static final String a = '';
 
+  // static final String a = '';
 }
-
-
-// static final String a = '';
