@@ -1,38 +1,36 @@
 class Assets {
-  static final String _base = 'assets/';
+  static const String _base = "assets/";
 
-  static final String _imgBase = 'assets/images/';
+  static const String _imgBase = "assets/images/";
 
-  static final String doclensewhite = "${_base}doclensewhite.svg";
+  static const String doclensewhite = "${_base}doclensewhite.svg";
 
-  static final String doclenselight = "${_imgBase}doclenselight.svg";
+  static const String doclenselight = "${_imgBase}doclenselight.svg";
 
-  static final String scanImg = "${_imgBase}scan.jpg";
+  static const String scanImg = "${_imgBase}scan.jpg";
 
-  static final String linkedinIcon =
-      'https://img.icons8.com/fluent/48/000000/linkedin-circled.png';
+  static const String linkedinIcon =
+      "https://img.icons8.com/fluent/48/000000/linkedin-circled.png";
 
-  static final String githubIcon =
-      'https://img.icons8.com/fluent/50/000000/github.png';
+  static const String githubIcon =
+      "https://img.icons8.com/fluent/50/000000/github.png";
 
-  static final String curveSvg = 'assets/aboutPage/curve.svg';
+  static const String curveSvg = "assets/aboutPage/curve.svg";
 
-  static final String curvelightSvg = 'assets/aboutPage/curvelight.svg';
+  static const String curvelightSvg = "assets/aboutPage/curvelight.svg";
 
-  static final String doclensewhiteSvg = 'assets/doclensewhite.svg';
+  static const String doclensewhiteSvg = "assets/doclensewhite.svg";
 
-  static final String doclenselightSvg = 'assets/images/doclenselight.svg';
+  static const String doclenselightSvg = "assets/images/doclenselight.svg";
 
-  static final String undrawSvg = 'assets/aboutPage/undraw_At_work_re_qotl.svg';
+  static const String undrawSvg = "assets/aboutPage/undraw_At_work_re_qotl.svg";
 
-  static final String undrawUpSvg =
-      'assets/aboutPage/undraw_Upload_re_pasx.svg';
+  static const String undrawUpSvg =
+      "assets/aboutPage/undraw_Upload_re_pasx.svg";
 
-  static final String doclensewhitesmall = "assets/doclensewhitesmall.svg";
+  static const String doclensewhitesmall = "assets/doclensewhitesmall.svg";
 
-  static final String doclenselightsmall = 'assets/doclenselightsmall.svg';
-
-// static final String a="";
+  static const String doclenselightsmall = "assets/doclenselightsmall.svg";
 
 // static final String a="";
 
@@ -40,7 +38,5 @@ class Assets {
 
 // static final String a="";
 
+// static final String a="";
 }
-
-
-// static final String a="";
