@@ -1,14 +1,14 @@
 class RouteConstants {
-  static const starredDocumentsScreen = '/starred-documents-screen';
-  static const settingsScreen = '/settings-screen';
-  static const aboutAppScreen = '/about-app-screen';
-  static const homeScreen = '/home-screen';
-  static const multiDelete = '/multi-delete';
-  static const imageView = '/image-view';
-  static const contactDeveloperScreen = '/contact-developer-screen';
-  static const photoFilterSelector = '/photo-filter-selector';
-  static const filterImageScreen = '/filter-image-screen';
-  static const pdfConversionScreen = '/pdf-conversion-screen';
-  static const pdfPreviewScreen = '/pdf-preview-screen';
-  static const folderPickerPage = '/folder-picker-page';
+  static const String starredDocumentsScreen = "/starred-documents-screen";
+  static const String settingsScreen = "/settings-screen";
+  static const String aboutAppScreen = "/about-app-screen";
+  static const String homeScreen = "/home-screen";
+  static const String multiDelete = "/multi-delete";
+  static const String imageView = "/image-view";
+  static const String contactDeveloperScreen = "/contact-developer-screen";
+  static const String photoFilterSelector = "/photo-filter-selector";
+  static const String filterImageScreen = "/filter-image-screen";
+  static const String pdfConversionScreen = "/pdf-conversion-screen";
+  static const String pdfPreviewScreen = "/pdf-preview-screen";
+  static const String folderPickerPage = "/folder-picker-page";
 }
