@@ -1,11 +1,9 @@
-import 'dart:io';
-import 'package:flutter/cupertino.dart';
+import "dart:io";
+import "package:flutter/cupertino.dart";
 
 class ImageList extends ChangeNotifier {
   List<File> imagelist = <File>[];
   List<String> imagepath = <String>[];
 
-  int length() {
-    return imagelist.length;
-  }
+  int length() => imagelist.length;
 }
