@@ -1,5 +1,7 @@
-import 'package:doclense/services/shared_preferences.dart';
-import 'package:flutter/material.dart';
+import "dart:async";
+
+import "package:doclense/services/shared_preferences.dart";
+import "package:flutter/material.dart";
 
 class DarkThemeProvider with ChangeNotifier {
   SharedPreferencesService darkThemePreference = SharedPreferencesService();
@@ -9,7 +11,12 @@ class DarkThemeProvider with ChangeNotifier {
 
   set darkTheme(bool? value) {
     _darkTheme = value ?? false;
-    darkThemePreference.setSharedPreferenceValue("themeMode", value);
+    unawaited(
+      darkThemePreference.setSharedPreferenceValue(
+        "themeMode",
+        value,
+      ),
+    );
     notifyListeners();
   }
 }
