@@ -1,23 +1,22 @@
-import 'dart:async';
-import 'dart:io';
+import "dart:async";
+import "dart:io";
 
-import 'package:doclense/configs/app_typography.dart';
-import 'package:doclense/constants/appstrings.dart';
-import 'package:flutter/material.dart';
-import 'package:pdf/pdf.dart';
-import 'package:printing/printing.dart';
-// import 'package:flutter_full_pdf_viewer/flutter_full_pdf_viewer.dart';
+import "package:doclense/configs/app_typography.dart";
+import "package:doclense/constants/appstrings.dart";
+import "package:flutter/material.dart";
+import "package:pdf/pdf.dart";
+import "package:printing/printing.dart";
 
 class PdfPreviewScreen extends StatefulWidget {
+  const PdfPreviewScreen({required this.path, required this.name, super.key});
   final String path;
   final String name;
-  const PdfPreviewScreen({required this.path, required this.name});
 
   @override
-  _PdfPreviewScreenState createState() => _PdfPreviewScreenState();
+  PdfPreviewScreenState createState() => PdfPreviewScreenState();
 }
 
-class _PdfPreviewScreenState extends State<PdfPreviewScreen> {
+class PdfPreviewScreenState extends State<PdfPreviewScreen> {
   @override
   void initState() {
     super.initState();
@@ -25,45 +24,36 @@ class _PdfPreviewScreenState extends State<PdfPreviewScreen> {
   }
 
   void homePageTimer() {
-    Timer(const Duration(), () {
-      Navigator.of(context).popUntil((route) => route.isFirst);
+    Timer(Duration.zero, () {
+      Navigator.of(context).popUntil((Route<dynamic> route) => route.isFirst);
     });
   }
 
   @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(
-          S.pdfPreview,
-          style: AppText.h4b,
+  Widget build(BuildContext context) => Scaffold(
+        appBar: AppBar(
+          title: Text(
+            S.pdfPreview,
+            style: AppText.h4b,
+          ),
         ),
-      ),
-      body: PdfPreview(
-        canDebug: false,
-        canChangeOrientation: false,
-        canChangePageFormat: false,
-        allowPrinting: false,
-        allowSharing: false,
-        pageFormats: <String, PdfPageFormat>{
-          'A4': PdfPageFormat.a4,
-          'Letter': PdfPageFormat.letter,
-          'A3': PdfPageFormat.a3,
-          'A5': PdfPageFormat.a5,
-          'Standard': PdfPageFormat.standard,
-        },
-        build: (PdfPageFormat format) {
-          return File(widget.path).readAsBytes();
-        },
-        // onPrinted: _showPrintedToast,
-        // onShared: _showSharedToast,
-      ),
-    );
-    // PDFViewerScaffold(
-    // path: widget.path,
-    // appBar: AppBar(
-    //   backgroundColor: Colors.blue,
-    // ),
-    // );
-  }
+        body: PdfPreview(
+          canDebug: false,
+          canChangeOrientation: false,
+          canChangePageFormat: false,
+          allowPrinting: false,
+          allowSharing: false,
+          pageFormats: const <String, PdfPageFormat>{
+            "A4": PdfPageFormat.a4,
+            "Letter": PdfPageFormat.letter,
+            "A3": PdfPageFormat.a3,
+            "A5": PdfPageFormat.a5,
+            "Standard": PdfPageFormat.standard,
+          },
+          build: (PdfPageFormat format) async =>
+              File(widget.path).readAsBytes(),
+          // onPrinted: _showPrintedToast,
+          // onShared: _showSharedToast,
+        ),
+      );
 }
