@@ -1,13 +1,12 @@
-import 'package:doclense/configs/app_typography.dart';
-import 'package:doclense/configs/space.dart';
-import 'package:flutter/material.dart';
+import "package:doclense/configs/app_typography.dart";
+import "package:doclense/configs/space.dart";
+import "package:flutter/material.dart";
 
 class SettingText extends StatelessWidget {
-  const SettingText({required this.text});
+  const SettingText({required this.text, super.key});
   final String text;
   @override
-  Widget build(BuildContext context) {
-    return Expanded(
+  Widget build(BuildContext context) => Expanded(
       flex: 5,
       child: Padding(
         padding: Space.all(),
@@ -17,5 +16,4 @@ class SettingText extends StatelessWidget {
         ),
       ),
     );
-  }
 }
