@@ -1,32 +1,36 @@
-import 'dart:io';
+// ignore_for_file: avoid_dynamic_calls
 
-import 'package:camera/camera.dart';
-import 'package:doclense/configs/app_dimensions.dart';
-import 'package:doclense/configs/app_typography.dart';
-import 'package:doclense/configs/space.dart';
-import 'package:doclense/constants/appstrings.dart';
-import 'package:doclense/ui_components/main_drawer.dart';
-import 'package:doclense/providers/theme_provider.dart';
-import 'package:doclense/ui_components/double_back_to_close_snackbar.dart';
-import 'package:doclense/utils/image_converter.dart' as image_converter;
-import 'package:double_back_to_close_app/double_back_to_close_app.dart';
-import 'package:flutter/material.dart';
-import 'package:hive_flutter/hive_flutter.dart';
-import 'package:open_file/open_file.dart';
-import 'package:provider/provider.dart';
-import 'package:share_plus/share_plus.dart';
+import "dart:io";
+
+import "package:doclense/configs/app_dimensions.dart";
+import "package:doclense/configs/app_typography.dart";
+import "package:doclense/configs/space.dart";
+import "package:doclense/constants/appstrings.dart";
+import "package:doclense/providers/theme_provider.dart";
+import "package:doclense/ui_components/double_back_to_close_snackbar.dart";
+import "package:doclense/ui_components/main_drawer.dart";
+import "package:doclense/utils/image_converter.dart" as image_converter;
+import "package:double_back_to_close_app/double_back_to_close_app.dart";
+import "package:flutter/material.dart";
+import "package:hive_flutter/hive_flutter.dart";
+import "package:open_file/open_file.dart";
+import "package:provider/provider.dart";
+import "package:share_plus/share_plus.dart";
 
 class Starred extends StatefulWidget {
+  const Starred({super.key});
+
   @override
-  _StarredState createState() => _StarredState();
+  StarredState createState() => StarredState();
 }
 
-class _StarredState extends State<Starred> {
+class StarredState extends State<Starred> {
   @override
   Widget build(BuildContext context) {
-    final themeChange = Provider.of<DarkThemeProvider>(context);
+    final DarkThemeProvider themeChange =
+        Provider.of<DarkThemeProvider>(context);
     return Scaffold(
-      drawer: MainDrawer(),
+      drawer: const MainDrawer(),
       appBar: AppBar(
         title: Text(
           S.starredDocs,
@@ -48,23 +52,27 @@ class _StarredState extends State<Starred> {
       ),
       body: DoubleBackToCloseApp(
         snackBar: doubleBackToCloseSnackBar(),
-        child: ValueListenableBuilder<Box>(
-          valueListenable: Hive.box('starred').listenable(),
-          builder: (context, starredBox, widget) {
-            if (starredBox.getAt(0).length == 0) {
+        child: ValueListenableBuilder<Box<dynamic>>(
+          valueListenable: Hive.box("starred").listenable(),
+          builder:
+              (BuildContext context, Box<dynamic> starredBox, Widget? widget) {
+            if ((starredBox.getAt(0) as List<dynamic>).isEmpty) {
               return const Center(
                 child: Text("No PDFs Starred Yet !! "),
               );
             }
             return ListView.builder(
-              itemCount: starredBox.getAt(0).length as int,
-              itemBuilder: (context, index) {
+              itemCount: (starredBox.getAt(0) as List<dynamic>).length,
+              itemBuilder: (BuildContext context, int index) {
                 final Image previewImage = image_converter.base64StringToImage(
-                    starredBox.getAt(0)[index][2] as String);
+                  (starredBox.getAt(0) as List<dynamic>)[index][2] as String,
+                );
                 return GestureDetector(
-                  onTap: () {
-                    print('tapped');
-                    OpenFile.open(starredBox.getAt(0)[index][0] as String);
+                  onTap: () async {
+                    debugPrint("tapped");
+                    await OpenFile.open(
+                      starredBox.getAt(0)[index][0] as String,
+                    );
                   },
                   child: Padding(
                     padding: Space.all(0.75),
@@ -74,7 +82,7 @@ class _StarredState extends State<Starred> {
                           ? Colors.grey[700]
                           : Colors.white,
                       child: Row(
-                        children: [
+                        children: <Widget>[
                           Container(
                             width: AppDimensions.width(25),
                             padding: Space.all(),
@@ -82,12 +90,12 @@ class _StarredState extends State<Starred> {
                           ),
                           Column(
                             crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
+                            children: <Widget>[
                               Padding(
                                 padding: Space.all(0.4),
                                 child: Text(
                                   (starredBox.getAt(0)[index][0] as String)
-                                      .split('/')
+                                      .split("/")
                                       .last
                                       .trim(),
                                   style: AppText.b1,
@@ -96,13 +104,13 @@ class _StarredState extends State<Starred> {
                               Padding(
                                 padding: Space.all(0.4),
                                 child: Text(
-                                  '${starredBox.getAt(0)[index][1]}',
+                                  "${starredBox.getAt(0)[index][1]}",
                                   style: AppText.l1,
                                 ),
                               ),
                               Space.y!,
                               Row(
-                                children: [
+                                children: <Widget>[
                                   SizedBox(
                                     width: (MediaQuery.of(context).size.width *
                                             MediaQuery.of(context)
@@ -110,32 +118,36 @@ class _StarredState extends State<Starred> {
                                         0.1,
                                   ),
                                   IconButton(
-                                      icon: Icon(
-                                        Icons.share,
-                                        color: themeChange.darkTheme
-                                            ? Colors.white70
-                                            : Colors.grey,
-                                        size: AppDimensions.font(8),
-                                      ),
-                                      onPressed: () async {
-                                        await _onShare(starredBox, index);
-                                      }),
+                                    icon: Icon(
+                                      Icons.share,
+                                      color: themeChange.darkTheme
+                                          ? Colors.white70
+                                          : Colors.grey,
+                                      size: AppDimensions.font(8),
+                                    ),
+                                    onPressed: () async {
+                                      await _onShare(starredBox, index);
+                                    },
+                                  ),
                                   IconButton(
-                                      icon: Icon(
-                                        Icons.star,
-                                        color: themeChange.darkTheme
-                                            ? Colors.white70
-                                            : Colors.grey,
-                                        size: AppDimensions.font(8),
-                                      ),
-                                      onPressed: () async {
-                                        _onStarred(index, context);
-                                      }),
+                                    icon: Icon(
+                                      Icons.star,
+                                      color: themeChange.darkTheme
+                                          ? Colors.white70
+                                          : Colors.grey,
+                                      size: AppDimensions.font(8),
+                                    ),
+                                    onPressed: () async {
+                                      await _onStarred(index, context);
+                                    },
+                                  ),
                                   IconButton(
-                                    icon: Icon(Icons.more_vert,
-                                        color: themeChange.darkTheme
-                                            ? Colors.white70
-                                            : Colors.grey),
+                                    icon: Icon(
+                                      Icons.more_vert,
+                                      color: themeChange.darkTheme
+                                          ? Colors.white70
+                                          : Colors.grey,
+                                    ),
                                     onPressed: () async {},
                                   ),
                                 ],
@@ -155,9 +167,9 @@ class _StarredState extends State<Starred> {
     );
   }
 
-  void _onStarred(int index, BuildContext context) {
+  Future<void> _onStarred(int index, BuildContext context) async {
     setState(() {
-      Hive.box('starred').getAt(0).removeAt(index);
+      Hive.box("starred").getAt(0).removeAt(index);
     });
     ScaffoldMessenger.of(context).showSnackBar(
       SnackBar(
@@ -172,12 +184,12 @@ class _StarredState extends State<Starred> {
   Future<void> _onShare(Box<dynamic> starredBox, int index) async {
     final File file = File(await starredBox.getAt(0)[index][0] as String);
 
-    final path = file.path;
+    final String path = file.path;
 
-    print(path);
+    debugPrint(path);
 
-    Share.shareXFiles(
-      [XFile(path)],
+    await Share.shareXFiles(
+      <XFile>[XFile(path)],
       text: S.yourPDF,
     );
   }
