@@ -1,3 +1,5 @@
-import 'package:flutter/material.dart';
+import "package:flutter/material.dart";
 
-SnackBar doubleBackToCloseSnackBar() => const SnackBar(content: Text('Press back again to close'),);
+SnackBar doubleBackToCloseSnackBar() => const SnackBar(
+      content: Text("Press back again to close"),
+    );
