@@ -1,22 +1,21 @@
-import 'package:doclense/configs/app_dimensions.dart';
-import 'package:doclense/configs/app_typography.dart';
-import 'package:flutter/material.dart';
+import "package:doclense/configs/app_dimensions.dart";
+import "package:doclense/configs/app_typography.dart";
+import "package:flutter/material.dart";
 
 class DrawerNavItem extends StatelessWidget {
-  final VoidCallback callback;
-  final IconData iconData;
-  final String navItemTitle;
 
   const DrawerNavItem({
     required this.callback,
     required this.iconData,
     required this.navItemTitle,
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
+  final VoidCallback callback;
+  final IconData iconData;
+  final String navItemTitle;
 
   @override
-  Widget build(BuildContext context) {
-    return ListTile(
+  Widget build(BuildContext context) => ListTile(
       onTap: callback,
       leading: Icon(iconData, size: AppDimensions.font(13)),
       title: Text(
@@ -24,5 +23,4 @@ class DrawerNavItem extends StatelessWidget {
         style: AppText.b2,
       ),
     );
-  }
 }
