@@ -1,29 +1,32 @@
-import 'package:doclense/configs/app_dimensions.dart';
-import 'package:doclense/configs/app_typography.dart';
-import 'package:doclense/configs/space.dart';
-import 'package:doclense/constants/appstrings.dart';
-import 'package:doclense/constants/assets.dart';
-import 'package:doclense/constants/route_constants.dart';
-import 'package:doclense/env.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
-import 'package:provider/provider.dart';
-import 'package:rating_dialog/rating_dialog.dart';
-import 'package:share_plus/share_plus.dart';
-import 'package:url_launcher/url_launcher.dart';
-import '../providers/theme_provider.dart';
-import 'drawer_nav_item.dart';
+import "package:doclense/configs/app_dimensions.dart";
+import "package:doclense/configs/app_typography.dart";
+import "package:doclense/configs/space.dart";
+import "package:doclense/constants/appstrings.dart";
+import "package:doclense/constants/assets.dart";
+import "package:doclense/constants/route_constants.dart";
+import "package:doclense/env.dart";
+import "package:doclense/providers/theme_provider.dart";
+import "package:doclense/ui_components/drawer_nav_item.dart";
+import "package:flutter/material.dart";
+import "package:flutter_svg/flutter_svg.dart";
+import "package:provider/provider.dart";
+import "package:rating_dialog/rating_dialog.dart";
+import "package:share_plus/share_plus.dart";
+import "package:url_launcher/url_launcher.dart";
 
 class MainDrawer extends StatelessWidget {
+  const MainDrawer({super.key});
+
   @override
   Widget build(BuildContext context) {
-    final themeChange = Provider.of<DarkThemeProvider>(context);
+    final DarkThemeProvider themeChange =
+        Provider.of<DarkThemeProvider>(context);
 
     return Drawer(
       child: SingleChildScrollView(
         child: Column(
           children: <Widget>[
-            Container(
+            ColoredBox(
               color: themeChange.darkTheme ? Colors.black : Colors.white10,
               child: Column(
                 children: <Widget>[
@@ -55,23 +58,24 @@ class MainDrawer extends StatelessWidget {
             DrawerNavItem(
               iconData: Icons.home,
               navItemTitle: S.home,
-              callback: () {
-                Navigator.of(context).pushReplacementNamed(
+              callback: () async {
+                await Navigator.of(context).pushReplacementNamed(
                   RouteConstants.homeScreen,
                 );
               },
             ),
             DrawerNavItem(
-                iconData: Icons.stars_rounded,
-                navItemTitle: S.starredDocuments,
-                callback: () {
-                  Navigator.of(context).pushReplacementNamed(
-                    RouteConstants.starredDocumentsScreen,
-                  );
-                }),
+              iconData: Icons.stars_rounded,
+              navItemTitle: S.starredDocuments,
+              callback: () async {
+                await Navigator.of(context).pushReplacementNamed(
+                  RouteConstants.starredDocumentsScreen,
+                );
+              },
+            ),
             DrawerNavItem(
-              callback: () {
-                Navigator.of(context).pushReplacementNamed(
+              callback: () async {
+                await Navigator.of(context).pushReplacementNamed(
                   RouteConstants.settingsScreen,
                 );
               },
@@ -84,8 +88,8 @@ class MainDrawer extends StatelessWidget {
             DrawerNavItem(
               iconData: Icons.info,
               navItemTitle: S.aboutApp,
-              callback: () {
-                Navigator.of(context).pushReplacementNamed(
+              callback: () async {
+                await Navigator.of(context).pushReplacementNamed(
                   RouteConstants.aboutAppScreen,
                 );
               },
@@ -93,34 +97,37 @@ class MainDrawer extends StatelessWidget {
             DrawerNavItem(
               navItemTitle: S.rateus,
               iconData: Icons.star_rate,
-              callback: () {
+              callback: () async {
                 Navigator.of(context).pop();
-                showDialog<void>(
-                    context: context,
-                    barrierDismissible: true,
-                    builder: (BuildContext context) {
-                      return RatingDialog(
-                        image: Padding(
-                          padding: EdgeInsets.fromLTRB(
-                              AppDimensions.normalize(10), 0, 0, 0),
-                          child: themeChange.darkTheme
-                              ? SvgPicture.asset(Assets.doclensewhiteSvg)
-                              : SvgPicture.asset(Assets.doclenselightSvg),
-                        ),
-                        title: Text(
-                          S.howWasExperience,
-                          textAlign: TextAlign.center,
-                        ),
-                        onSubmitted: (RatingDialogResponse) {
-                          _launchURL();
-                        },
-                        submitButtonText: S.submit,
-                        message: Text(
-                          S.letUsKnow,
-                          textAlign: TextAlign.center,
-                        ),
-                      );
-                    });
+                await showDialog<void>(
+                  context: context,
+                  builder: (BuildContext context) => RatingDialog(
+                    image: Padding(
+                      padding: EdgeInsets.fromLTRB(
+                        AppDimensions.normalize(10),
+                        0,
+                        0,
+                        0,
+                      ),
+                      child: themeChange.darkTheme
+                          ? SvgPicture.asset(Assets.doclensewhiteSvg)
+                          : SvgPicture.asset(Assets.doclenselightSvg),
+                    ),
+                    title: const Text(
+                      S.howWasExperience,
+                      textAlign: TextAlign.center,
+                    ),
+                    onSubmitted:
+                        (RatingDialogResponse ratingDialogResponse) async {
+                      await _launchURL();
+                    },
+                    submitButtonText: S.submit,
+                    message: const Text(
+                      S.letUsKnow,
+                      textAlign: TextAlign.center,
+                    ),
+                  ),
+                );
               },
             ),
             Divider(
@@ -129,8 +136,8 @@ class MainDrawer extends StatelessWidget {
             DrawerNavItem(
               navItemTitle: S.shareTheApp,
               iconData: Icons.share,
-              callback: () {
-                Share.share(S.shareMessage, subject: Env.appname);
+              callback: () async {
+                await Share.share(S.shareMessage, subject: Env.appname);
               },
             ),
             // ListTile(
@@ -152,11 +159,11 @@ class MainDrawer extends StatelessWidget {
 }
 
 Future<void> _launchURL() async {
-  const url =
-      'https://github.com/smaranjitghose/DocLense'; //!paste link of app once uploaded on play store
+  const String url = "https://github.com/smaranjitghose/DocLense";
+  //!paste link of app once uploaded on play store
   if (await canLaunchUrl(Uri.parse(url))) {
     await launchUrl(Uri.parse(url));
   } else {
-    throw 'Could not launch $url';
+    debugPrint("Could not launch $url");
   }
 }
